api/requirements.txt
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Caching==1.10.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.25.0
mysqlclient==2.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==1.4.2
reportlab==3.5.67
requests==2.27.1
six==1.15.0
SQLAlchemy==1.4.12
urllib3==1.26.5
Werkzeug==1.0.1
